Changed some prototypes to use size_t.

+++ snmp/g_utils.c
@@ -173,7 +173,7 @@ gsnmp_identity_get_label(GSnmpIdentity const *table,
 
 int
 gsnmp_attr_assign(GSList *vbl,
-		  guint32 const *base, guint8 const len,
+		  guint32 const *base, size_t const len,
 		  const GSnmpAttribute *attributes, const gpointer p)
 {
     GSList *elem;
@@ -324,7 +324,7 @@ gsnmp_attr_assign(GSList *vbl,
 
 void
 gsnmp_attr_get(const GSnmpSession *s, GSList **vbl,
-	       guint32 *base, guint8 const len, guint const idx,
+	       guint32 *base, size_t const len, guint const idx,
 	       const GSnmpAttribute *attributes, const gint mask)
 {
     int i;
@@ -344,7 +344,7 @@ gsnmp_attr_get(const GSnmpSession *s, GSList **vbl,
 
 void
 gsnmp_attr_set(const GSnmpSession *s, GSList **vbl,
-	       guint32 *base, guint8 const len, guint const idx,
+	       guint32 *base, size_t const len, guint const idx,
 	       const GSnmpAttribute *attributes, const gint mask,
 	       const gpointer p)
 {
